feat: use save hook
